Add matrix_scale project - feedback matrix tuned to MIDI
The idea is to have a bank of resonators tuned so that using different
combinations of them in sequence results in feedback at a wide variety
of pitches.

The tunings of the resonators are 2f, 3f, 4f, 5f, and 8f. Thes
combinations produce a largish number of related pitches and a few
unrelated ones. I've mapped where these pitches lie on the keyboard and
plan to use these as a lookup table to identify which sets of resonators
to use for each note.

Hopefully the intended pitches will appear when playing monophonically,
but the result of playing chords will be chaotic and hopefully fun.