Add clippedResonator
Should be useful if we want a more plucked sound - it's safe from
exploding amplitude, but linear in the maximum range.

I tried some experiments with using Limiter as the delay. Unfortunately
the pitches were all wrong. I think this might be because the Limiter
was oscillating at audio frequency, because of the very short delay
time. It could be a fun thing to explore at some point.